Instrument (un)wrapInternal only for the latest stable version of the coroutine fork
Plugins may depend on earlier versions of IC/IU, and being worked on in the latest IDE with the latest debugger agent results in incompatibility and thus unexpected CCEs.

Capture/insertion points will still be instrumented, but in older versions both of the insertion points do not exist, which basically means async stack traces for flows won't work for older versions. Which is still better than breaking the execution with unexpected exceptions.